refactor(permission): 퍼미션 확인 객체 분리
