+++ app/src/main/java/woowacourse/movie/util/permission/Permissions.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.util
+package woowacourse.movie.util.permission
 
 import android.app.Activity
 import android.content.Context
@@ -18,14 +18,5 @@ fun requestPermission(
     }
 }
 
-// fun Context.hasPermission(permission: String): Boolean =
-//    ContextCompat.checkSelfPermission(this, permission) == PackageManager.PERMISSION_GRANTED
-
-/*
-    SettingFragment에서 사용중인 메서드
-    requireContext를 사용할 때에는 원하는대로 작동하지 않음 (현재 앱에 대한 permission의 여부에 따른 변화 적용x)
-    requireActivity를 사용하였을 때에는 원하는대로 작동.
-    ...?
- */
 fun hasPermission(context: Context, permission: String): Boolean =
     ContextCompat.checkSelfPermission(context, permission) == PackageManager.PERMISSION_GRANTED
